requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
PyJWT==2.3.0
redis==3.5.3
requests==2.26.0
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
